+++ src/pretalx/orga/views/submission.py
@@ -292,6 +292,10 @@ class SubmissionContent(ActionFromUrl, SubmissionViewMixin, CreateOrUpdateView):
             target='submission',
             submission=submission,
             event=self.request.event,
+            for_reviewers=(
+                not self.request.user.has_perm('orga.change_submissions', self.request.event)
+                and self.request.user.has_perm('orga.view_review_dashboard', self.request.event)
+            ),
         )
 
     @context